+++ docs/make.jl
@@ -1,3 +1,6 @@
+# Accept PalmerPenguins download on CI
+ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
+
 using Pkg: Pkg
 Pkg.develop(Pkg.PackageSpec(; path = dirname(@__DIR__)))
 
@@ -19,6 +22,9 @@ using DynamicPPL: DynamicPPL
 using MCMCChains: MCMCChains
 using PairPlots: PairPlots
 using Turing: Turing
+using PalmerPenguins: PalmerPenguins
+using DataFrames: DataFrames
+
 
 links = InterLinks(
     "AdvancedHMC" => "https://turinglang.org/AdvancedHMC.jl/stable/",
@@ -52,7 +58,13 @@ ENV["GKSwstype"] = "100"
 
 makedocs(;
     sitename = "FlexiChains.jl",
-    format = Documenter.HTML(; assets = ["assets/custom.css"]),
+    format = Documenter.HTML(;
+        # Some pictures are larger than the default threshold of 8 KB
+        example_size_threshold = 12 * 2^10, # 12 KB
+        # And some pages are larger than the default threshold of 100 KB
+        size_threshold_warn = 200 * 2^10, # 200 KB
+        assets = ["assets/custom.css"],
+    ),
     modules = modules,
     pages = [
         "index.md",
@@ -72,7 +84,7 @@ makedocs(;
         "migration.md",
     ],
     checkdocs = :exports,
-    warnonly = true,
+    warnonly = false,
     doctest = false,
     plugins = [links],
 )